# Overview
This is to finish removal of Offers as well as new work to add form fields and send input(s) via view model out for processing. This does not include crossing over the boundary into the DTO - this will be dependent on IES-841.

# Testing
![2022-02-17_14-56-06.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/785/attachments/2022-02-17_14-56-06.png)

![2022-02-17_14-58-40.png](https://idetfs01.us.lmco.com/Estimating/ec33f0ce-06e6-4674-9d97-643c8eb6e517/_apis/git/repositories/c962790d-14a5-4524-9fe8-35b4cce33cc3/pullRequests/785/attachments/2022-02-17_14-58-40.png)

# Known Issues / Comments

This does not currently save the new fields. On clicking save, it sends them to the controller which adds them to the view model, but the mechanism to transfer the data to the DTO - then on to the actual DB persistence - is not currently in place.

This also updates the `PtmContracts` branch with UAT so you will see a few extra changes related to exports.

+++ GenBOE/GenTRAC.Tests/ActionLogic/ContractsControllerLogicTest.cs
@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright company="Lockheed Martin Corporation">
-//     Copyright (c) 2011 - 2022 Lockheed Martin Corporation
+//     Copyright (c) 2011 - 2021 Lockheed Martin Corporation
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -177,7 +177,7 @@ namespace GenTRAC.Tests.ActionLogic
             Assert.AreEqual(contractDto.NegotiationsSubmitted, contractsModelView.NegotiationsSubmittedDt);
             Assert.AreEqual(contractDto.PreviouslySubmittedROM.ToString(), contractsModelView.PreviouslySubmittedRoms.ElementAtOrDefault(0).Value);
             Assert.AreEqual(contractDto.PreviouslySubmittedROM, contractsModelView.PreviouslySubmittedROM);
-            Assert.AreEqual(contractDto.EppDelegationAuthority, contractsModelView.EppDelegationAuthority);
+            Assert.AreEqual(contractDto.EppDelegationAuthority, (int)contractsModelView.EppDelegationAuthority);
             Assert.AreEqual(contractDto.ProgramEppDate, contractsModelView.ProgramEppDate);
             Assert.AreEqual(contractDto.LobEppDate, contractsModelView.LobEppDate);
             Assert.AreEqual(contractDto.PreSpaceEppDate, contractsModelView.PreSpaceEppDate);
@@ -232,8 +232,8 @@ namespace GenTRAC.Tests.ActionLogic
         {
             // Create new Contracts Controller Logic object
             ContractsControllerLogic sut = this.CreateSystem();
-            
-            // TODO: SaveContract unit test. 
+
+            // TODO: SaveContract unit test.
             
             // Assert that the id return from the save is the one we expect.
             // Verify method - that it ran, but if we get an Id, may not be needed)
